Create full test users when local_only is relevant
When using Fabricate(:account), the account's user field is set to nil.
PostStatusService attempts to access this with
`@account.user&.setting_default_federation`, and thus treats a nil user
as defaulting to local-only posts. If we instead create a test user, and
then use its account in the test, the user's settings are set to their
default values, which include `setting_default_federation: true`, so
PostStatusService will federate the test user's posts, and these tests
can operate as before.