Added limiter for invites
A user can now only issue 3 invites / week (the 3 is arbitrary).
The behaviour is as expected, but the visualisation is not great:
instead of displaying a nice modal (similar to how rate limiter
errors are displayed for statuses), the user is redirected to
a HTTP 429 ERROR page (similar to how 404 errors are displayed).